refactor(DUP): time based actions
Moves `execute_time_based_actions/2` in `TimeBasedActions` module, totally
removing time based actions from impl.
